Add basic script that plots a visible satellite image with goes2go
This script should be seen as a minimal working example. The plotting routine still needs to be adapted to the structure created in goes.py and should probably be integrated into the latter goes.py script, either replacing the current way of plotting goes images, or as an alternative option.